(PC-17596)[PRO] chore: refactor getEducationalCategoriesAdapter
