+++ pro/src/screens/OfferEducational/__specs__/OfferEducationalEventAddressStep.spec.tsx
@@ -25,10 +25,12 @@ describe('screens | OfferEducational : event address step', () => {
     beforeEach(() => {
       props = {
         ...defaultCreationProps,
-        educationalCategories: categoriesFactory([{ id: 'CAT_1' }]),
-        educationalSubCategories: subCategoriesFactory([
-          { categoryId: 'CAT_1', id: 'SUBCAT_1' },
-        ]),
+        categories: {
+          educationalCategories: categoriesFactory([{ id: 'CAT_1' }]),
+          educationalSubCategories: subCategoriesFactory([
+            { categoryId: 'CAT_1', id: 'SUBCAT_1' },
+          ]),
+        },
       }
       store = {
         features: {
@@ -119,10 +121,12 @@ describe('screens | OfferEducational : event address step', () => {
             ],
           },
         ]),
-        educationalCategories: categoriesFactory([{ id: 'CAT_1' }]),
-        educationalSubCategories: subCategoriesFactory([
-          { categoryId: 'CAT_1', id: 'SUBCAT_1' },
-        ]),
+        categories: {
+          educationalCategories: categoriesFactory([{ id: 'CAT_1' }]),
+          educationalSubCategories: subCategoriesFactory([
+            { categoryId: 'CAT_1', id: 'SUBCAT_1' },
+          ]),
+        },
       }
     })
